fix(dashboard): address PR review — ARIA tabs, server page, query limits, shared utils
