+++ app/(dashboard)/dashboard/_components/recognition-feed-widget.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import { RecognitionFeed } from "../recognition/_components/recognition-feed";
 
 export function RecognitionFeedWidget() {
