+++ app/api/recognition/route.ts
@@ -45,6 +45,7 @@ export async function GET(request: NextRequest) {
 				},
 			},
 			orderBy: { createdAt: "desc" },
+			take: 50,
 		});
 
 		return Response.json({ success: true, data: cards });
